requirements.txt
streamlit==1.29.0
langchain==0.0.350
openai==1.12.0
qdrant-client==1.7.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.24.3
sentence-transformers==2.2.2
pydantic==2.4.2
requests==2.31.0 